requirements.txt
kafka-python>=2.0.2
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0

